[WIP] Enhance Graph to allow chart edits
The chart is by default a table if a curve is not detected. The columns
of the table can be chosen using a dropdown, the data can be transposed
and the chart type can be modified in-place. However, there is still a
bug that after the first message in the chat the chart type cannot be
modified.