enh: Update ChatbotWebsocketStreaming to store and display the last user and bot messages
This commit adds state variables to store the last user and bot messages in the ChatbotWebsocketStreaming component. It also updates the logic to set the last user and bot messages when a new message is sent or received. This will allow for easier access and display of the last messages in the chatbot interface.